Switched long,lat to lat,long globally.
...and definitively discontinuated utm2geo() because there is no way to
mantain reasonably backward compatibility paradigm because utm2geo_dec(), which
may be the most useful function of the original version (because
provides coordinates in the format expected by Google Maps APIs), has
also inverted its result order.

For this reason, I think the best solution is to definitively remove it
due tu force an evident fail in old code and force to update it.